feat: update members page layout
