+++ src/app/[locale]/admin/members/[id]/page.tsx
@@ -10,8 +10,10 @@ const NotFound = () => {
   const t = useTranslations()
 
   return (
-    <div className="rounded-lg bg-white p-6 text-center shadow-sm">
-      <p className="text-gray-600">{t('admin.members.notFound')}</p>
+    <div className="mx-auto max-w-(--breakpoint-xl) p-4 md:p-6">
+      <div className="rounded-lg bg-white p-6 text-center shadow-sm">
+        <p className="text-gray-600">{t('admin.members.notFound')}</p>
+      </div>
     </div>
   )
 }
@@ -38,7 +40,11 @@ const MemberViewPage = () => {
     return <NotFound />
   }
 
-  return <MemberView member={member} />
+  return (
+    <div className="mx-auto max-w-(--breakpoint-xl) p-4 md:p-6">
+      <MemberView member={member} />
+    </div>
+  )
 }
 
 export default MemberViewPage
