+++ src/app/[locale]/admin/members/layout.tsx
@@ -2,8 +2,6 @@ type LayoutProps = {
   children: React.ReactNode
 }
 
-const Layout = ({ children }: Readonly<LayoutProps>) => {
-  return <div className="mx-auto max-w-(--breakpoint-xl) p-4 md:p-6">{children}</div>
-}
+const Layout = ({ children }: Readonly<LayoutProps>) => <>{children}</>
 
 export default Layout
