+++ src/app/[locale]/admin/members/new/page.tsx
@@ -33,7 +33,7 @@ const NewMemberPage = () => {
   }
 
   return (
-    <>
+    <div className="mx-auto max-w-(--breakpoint-xl) p-4 md:p-6">
       <MemberForm
         initialFormData={getInitialFormData(null)}
         onCancel={handleCancel}
@@ -44,7 +44,7 @@ const NewMemberPage = () => {
         member={createdMember}
         onClose={handleDialogClose}
       />
-    </>
+    </div>
   )
 }
