+++ src/app/[locale]/admin/members/page.tsx
@@ -1,19 +1,12 @@
-'use client'
-
-import { MembersList } from '@components/features/admin/Members/MembersList'
+import { Suspense } from 'react'
 import { Spinner } from '@components/ui'
-import { useMembersQuery } from '@data/hooks/members'
-import { useTranslations } from 'next-intl'
 
-const MembersPage = () => {
-  const t = useTranslations()
-  const { data: members, isPending } = useMembersQuery()
+import MembersContainer from './MembersContainer'
 
-  if (isPending) return <Spinner label={t('common.labels.wait')} size="lg" />
-
-  if (!members) return null
-
-  return <MembersList members={members} />
-}
+const MembersPage = () => (
+  <Suspense fallback={<Spinner size="lg" />}>
+    <MembersContainer />
+  </Suspense>
+)
 
 export default MembersPage
