+++ src/components/features/admin/Members/MembersList/index.ts
@@ -1,1 +1,2 @@
+export { default as MemberFilters } from './MemberFilters'
 export { default as MembersList } from './MembersList'
